Move State to Migration\State
