+++ src/Migrator.php
@@ -16,6 +16,7 @@ use Spiral\Database\DatabaseManager;
 use Spiral\Database\Table;
 use Spiral\Migrations\Config\MigrationConfig;
 use Spiral\Migrations\Exception\MigrationException;
+use Spiral\Migrations\Migration\State;
 use Spiral\Migrations\MigrationInterface;
 use Spiral\Migrations\Migrator\MigrationsTable;
